+++ rxlifecycle-android/gradle.properties
@@ -1,0 +1,4 @@
+POM_NAME=RxLifecycle-Android
+POM_DESCRIPTION=RxLifecycle-Android
+POM_ARTIFACT_ID=rxlifecycle-android
+POM_PACKAGING=aar
