+++ rxlifecycle-android/src/main/java/com/trello/rxlifecycle/android/FragmentEvent.java
@@ -12,18 +12,22 @@
  * limitations under the License.
  */
 
-package com.trello.rxlifecycle;
+package com.trello.rxlifecycle.android;
 
 /**
- * Lifecycle events that can be emitted by Activities.
+ * Lifecycle events that can be emitted by Fragments.
  */
-public enum ActivityEvent {
+public enum FragmentEvent {
 
+    ATTACH,
     CREATE,
+    CREATE_VIEW,
     START,
     RESUME,
     PAUSE,
     STOP,
-    DESTROY
+    DESTROY_VIEW,
+    DESTROY,
+    DETACH
 
 }
