+++ rxlifecycle-android/src/test/java/com/trello/rxlifecycle/TestUtil.java
