+++ rxlifecycle-components/src/test/java/com/trello/rxlifecycle/components/RxFragmentLifecycleTest.java
@@ -15,8 +15,8 @@
 package com.trello.rxlifecycle.components;
 
 import android.app.Fragment;
-import com.trello.rxlifecycle.FragmentEvent;
 import com.trello.rxlifecycle.LifecycleProvider;
+import com.trello.rxlifecycle.android.FragmentEvent;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
