+++ rxlifecycle-navi/src/main/java/com/trello/rxlifecycle/navi/NaviLifecycle.java
@@ -3,9 +3,9 @@ package com.trello.rxlifecycle.navi;
 import android.support.annotation.CheckResult;
 import android.support.annotation.NonNull;
 import com.trello.navi.NaviComponent;
-import com.trello.rxlifecycle.ActivityEvent;
-import com.trello.rxlifecycle.FragmentEvent;
 import com.trello.rxlifecycle.LifecycleProvider;
+import com.trello.rxlifecycle.android.ActivityEvent;
+import com.trello.rxlifecycle.android.FragmentEvent;
 
 import static com.trello.rxlifecycle.internal.Preconditions.checkNotNull;
