+++ rxlifecycle-navi/src/test/java/com/trello/rxlifecycle/components/NaviFragmentLifecycleTest.java
@@ -17,8 +17,8 @@ package com.trello.rxlifecycle.components;
 import com.trello.navi.Event;
 import com.trello.navi.NaviComponent;
 import com.trello.navi.internal.NaviEmitter;
-import com.trello.rxlifecycle.FragmentEvent;
 import com.trello.rxlifecycle.LifecycleProvider;
+import com.trello.rxlifecycle.android.FragmentEvent;
 import com.trello.rxlifecycle.navi.NaviLifecycle;
 import org.junit.Test;
 import rx.Observable;
