+++ rxlifecycle/src/main/java/com/trello/rxlifecycle/RxLifecycle.java
@@ -16,8 +16,6 @@ package com.trello.rxlifecycle;
 
 import android.support.annotation.CheckResult;
 import android.support.annotation.NonNull;
-import android.view.View;
-import com.jakewharton.rxbinding.view.RxView;
 import rx.Observable;
 import rx.functions.Func1;
 
@@ -27,32 +25,6 @@ public class RxLifecycle {
 
     private RxLifecycle() {
         throw new AssertionError("No instances");
-    }
-
-    /**
-     * Deprecated and will be removed in a future release.
-     *
-     * Use {@link RxLifecycle#bindUntilEvent(Observable, Object)} instead, which does exactly the same thing.
-     */
-    @Deprecated
-    @NonNull
-    @CheckResult
-    public static <T> LifecycleTransformer<T> bindUntilFragmentEvent(
-        @NonNull final Observable<FragmentEvent> lifecycle, @NonNull final FragmentEvent event) {
-        return bindUntilEvent(lifecycle, event);
-    }
-
-    /**
-     * Deprecated and will be removed in a future release.
-     *
-     * Use {@link RxLifecycle#bindUntilEvent(Observable, Object)} instead, which does exactly the same thing.
-     */
-    @Deprecated
-    @NonNull
-    @CheckResult
-    public static <T> LifecycleTransformer<T> bindUntilActivityEvent(
-        @NonNull final Observable<ActivityEvent> lifecycle, @NonNull final ActivityEvent event) {
-        return bindUntilEvent(lifecycle, event);
     }
 
     /**
@@ -75,88 +47,6 @@ public class RxLifecycle {
         checkNotNull(event, "event == null");
 
         return new UntilEventObservableTransformer<>(lifecycle, event);
-    }
-
-    /**
-     * Binds the given source to an Activity lifecycle.
-     * <p>
-     * Use with {@link Observable#compose(Observable.Transformer)}:
-     * {@code source.compose(RxLifecycle.bindActivity(lifecycle)).subscribe()}
-     * <p>
-     * This helper automatically determines (based on the lifecycle sequence itself) when the source
-     * should stop emitting items. In the case that the lifecycle sequence is in the
-     * creation phase (CREATE, START, etc) it will choose the equivalent destructive phase (DESTROY,
-     * STOP, etc). If used in the destructive phase, the notifications will cease at the next event;
-     * for example, if used in PAUSE, it will unsubscribe in STOP.
-     * <p>
-     * Due to the differences between the Activity and Fragment lifecycles, this method should only
-     * be used for an Activity lifecycle.
-     *
-     * @param lifecycle the lifecycle sequence of an Activity
-     * * @return a reusable {@link Observable.Transformer} that unsubscribes the source during the Activity lifecycle
-     */
-    @NonNull
-    @CheckResult
-    public static <T> LifecycleTransformer<T> bindActivity(@NonNull final Observable<ActivityEvent> lifecycle) {
-        return bind(lifecycle, ACTIVITY_LIFECYCLE);
-    }
-
-    /**
-     * Binds the given source to a Fragment lifecycle.
-     * <p>
-     * Use with {@link Observable#compose(Observable.Transformer)}:
-     * {@code source.compose(RxLifecycle.bindFragment(lifecycle)).subscribe()}
-     * <p>
-     * This helper automatically determines (based on the lifecycle sequence itself) when the source
-     * should stop emitting items. In the case that the lifecycle sequence is in the
-     * creation phase (CREATE, START, etc) it will choose the equivalent destructive phase (DESTROY,
-     * STOP, etc). If used in the destructive phase, the notifications will cease at the next event;
-     * for example, if used in PAUSE, it will unsubscribe in STOP.
-     * <p>
-     * Due to the differences between the Activity and Fragment lifecycles, this method should only
-     * be used for a Fragment lifecycle.
-     *
-     * @param lifecycle the lifecycle sequence of a Fragment
-     * @return a reusable {@link Observable.Transformer} that unsubscribes the source during the Fragment lifecycle
-     */
-    @NonNull
-    @CheckResult
-    public static <T> LifecycleTransformer<T> bindFragment(@NonNull final Observable<FragmentEvent> lifecycle) {
-        return bind(lifecycle, FRAGMENT_LIFECYCLE);
-    }
-
-    /**
-     * Binds the given source to a View lifecycle.
-     * <p>
-     * Specifically, when the View detaches from the window, the sequence will be completed.
-     * <p>
-     * Use with {@link Observable#compose(Observable.Transformer)}:
-     * {@code source.compose(RxLifecycle.bindView(lifecycle)).subscribe()}
-     * <p>
-     * Warning: you should make sure to use the returned Transformer on the main thread,
-     * since we're binding to a View (which only allows binding on the main thread).
-     *
-     * @param view the view to bind the source sequence to
-     * @return a reusable {@link Observable.Transformer} that unsubscribes the source during the View lifecycle
-     */
-    @NonNull
-    @CheckResult
-    public static <T> LifecycleTransformer<T> bindView(@NonNull final View view) {
-        checkNotNull(view, "view == null");
-
-        return bind(RxView.detaches(view));
-    }
-
-    /**
-     * Deprecated and will be removed in a future release.
-     *
-     * Use {@link RxLifecycle#bind(Observable)} instead, which does exactly the same thing.
-     */
-    @Deprecated
-    @NonNull
-    @CheckResult
-    public static <T, E> LifecycleTransformer<T> bindView(@NonNull final Observable<? extends E> lifecycle) {
-        return bind(lifecycle);
     }
 
     /**
@@ -204,61 +94,4 @@ public class RxLifecycle {
         // Keep emitting from source until the corresponding event occurs in the lifecycle
         return new UntilCorrespondingEventObservableTransformer<>(lifecycle.share(), correspondingEvents);
     }
-
-    // Figures out which corresponding next lifecycle event in which to unsubscribe, for Activities
-    private static final Func1<ActivityEvent, ActivityEvent> ACTIVITY_LIFECYCLE =
-        new Func1<ActivityEvent, ActivityEvent>() {
-            @Override
-            public ActivityEvent call(ActivityEvent lastEvent) {
-                switch (lastEvent) {
-                    case CREATE:
-                        return ActivityEvent.DESTROY;
-                    case START:
-                        return ActivityEvent.STOP;
-                    case RESUME:
-                        return ActivityEvent.PAUSE;
-                    case PAUSE:
-                        return ActivityEvent.STOP;
-                    case STOP:
-                        return ActivityEvent.DESTROY;
-                    case DESTROY:
-                        throw new OutsideLifecycleException("Cannot bind to Activity lifecycle when outside of it.");
-                    default:
-                        throw new UnsupportedOperationException("Binding to " + lastEvent + " not yet implemented");
-                }
-            }
-        };
-
-    // Figures out which corresponding next lifecycle event in which to unsubscribe, for Fragments
-    private static final Func1<FragmentEvent, FragmentEvent> FRAGMENT_LIFECYCLE =
-        new Func1<FragmentEvent, FragmentEvent>() {
-            @Override
-            public FragmentEvent call(FragmentEvent lastEvent) {
-                switch (lastEvent) {
-                    case ATTACH:
-                        return FragmentEvent.DETACH;
-                    case CREATE:
-                        return FragmentEvent.DESTROY;
-                    case CREATE_VIEW:
-                        return FragmentEvent.DESTROY_VIEW;
-                    case START:
-                        return FragmentEvent.STOP;
-                    case RESUME:
-                        return FragmentEvent.PAUSE;
-                    case PAUSE:
-                        return FragmentEvent.STOP;
-                    case STOP:
-                        return FragmentEvent.DESTROY_VIEW;
-                    case DESTROY_VIEW:
-                        return FragmentEvent.DESTROY;
-                    case DESTROY:
-                        return FragmentEvent.DETACH;
-                    case DETACH:
-                        throw new OutsideLifecycleException("Cannot bind to Fragment lifecycle when outside of it.");
-                    default:
-                        throw new UnsupportedOperationException("Binding to " + lastEvent + " not yet implemented");
-                }
-            }
-        };
-
 }
