+++ settings.gradle
@@ -1,4 +1,5 @@
 include ':rxlifecycle'
+include ':rxlifecycle-android'
 include ':rxlifecycle-components'
 include ':rxlifecycle-navi'
 include ':rxlifecycle-kotlin'
